v0 support for MongoDB (basic RAG)
