+++ phi/vectordb/mongodb/__init__.py
@@ -1,0 +1,3 @@
+from phi.vectordb.mongodb.mongodb import MDBVector
+
+__all__ = ["MDBVector"]
